mediauth/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
openai==1.3.5
chromadb==0.4.18
python-dotenv==1.0.0